requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
google-cloud-storage
google-cloud-pubsub
